+++ AbraCADabra/ObjectManagers/TorusManager.cs
@@ -54,8 +54,8 @@ namespace AbraCADabra
             return (Math.Sin(val), Math.Cos(val));
         }
 
-        public float UScale => 1;
-        public float VScale => 1;
+        public int UScale => 1;
+        public int VScale => 1;
 
         public bool IsUVValid(float u, float v) => true;
